changes config variables to uint16_t

+++ wled00/wled00.ino
@@ -244,11 +244,11 @@ bool aOtaEnabled = true;                      //ArduinoOTA allows easy updates d
 uint16_t userVar0 = 0, userVar1 = 0;
 
 //dmx CONFIG
-int DMXChannels = 7;                          // number of channels per fixture
-int DMXFixtureMap[15] = { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0};
+uint16_t DMXChannels = 7;                          // number of channels per fixture
+uint16_t DMXFixtureMap[15] = { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0};
                                               // assigns the different channels to different functions. See wled21_dmx.ino for more information.
-int DMXGap = 10;                              // gap between the fixtures. makes addressing easier because you don't have to memorize odd numbers when climbing up onto a rig.
-int DMXStart = 10;                            // start address of the first fixture
+uint16_t DMXGap = 10;                              // gap between the fixtures. makes addressing easier because you don't have to memorize odd numbers when climbing up onto a rig.
+uint16_t DMXStart = 10;                            // start address of the first fixture
